move dyn_expr
